Return Float from averagePower(forChannel:) to match AVFoundation
AVFoundation declares both power accessors as returning `float`:

    - (float)peakPowerForChannel:(NSUInteger)channelNumber;
    - (float)averagePowerForChannel:(NSUInteger)channelNumber;

Returning `Double` on Android meant that cross-platform code compiling
against the iOS signature failed to compile once transpiled to Kotlin.

+++ Sources/SkipAV/AVAudioRecorder.swift
@@ -156,9 +156,9 @@ public class AVAudioRecorder: KotlinConverting<MediaRecorder?> {
         return Float(recorder?.maxAmplitude ?? 0) / Float(32767.0)
     }
 
-    public func averagePower(forChannel channelNumber: Int) -> Double {
+    public func averagePower(forChannel channelNumber: Int) -> Float {
         // Android doesn't provide average power, so we'll return peak power
-        return Double(recorder?.maxAmplitude ?? 0) / 32767.0
+        return Float(recorder?.maxAmplitude ?? 0) / Float(32767.0)
     }
 }
 #endif
